+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- [github.com/InVisionApp/go-logger](https://github.com/InVisionApp/go-logger) integration
+
 ### Changed
 
 - Renamed `logrusshim` to `logrusadapter`